Revert rename changes directory->peer_directory
